Updated Emulator & Assembler + Added Programs
Updated the assembler with the rest of the functions and fixed some bugs. Also fixed bugs in the assembler and finished the display and output stuff. Added a bunch of new programs which can be converted to binary using the assembler and run using the emulator (not perfectly accurate to the computer being developed but close enough for what's being done currently).